+++ Runtime/Gameplay/ObjectEventChannelSO.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+using UnityEngine.Events;
+
+namespace Slax.EventSystem
+{
+    [CreateAssetMenu(menuName = "Events/Gameplay/Object Event Channel")]
+    public class ObjectEventChannelSO : EventChannelSO
+    {
+        public UnityAction<Object> OnEventRaised;
+
+        public void RaiseEvent(Object obj)
+        {
+            if (OnEventRaised != null)
+                OnEventRaised.Invoke(obj);
+        }
+    }
+
+}