+++ Runtime/Gameplay/SpriteEventChannelSO.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+using UnityEngine.Events;
+
+namespace Slax.EventSystem
+{
+    [CreateAssetMenu(menuName = "Events/Gameplay/Sprite Event Channel")]
+    public class SpriteEventChannelSO : EventChannelSO
+    {
+        public UnityAction<Sprite> OnEventRaised;
+
+        public void RaiseEvent(Sprite sprite)
+        {
+            if (OnEventRaised != null)
+                OnEventRaised.Invoke(sprite);
+        }
+    }
+
+}