+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2024-07-21
+### Added
+- Custom editor view to find all references of an event in loaded scenes
+- Custom editor window to create new custom event channels easily accessible from "Slax/Event System/Event Channel Creator"
+### New pre-supported event channels
+- AudioClip
+- Collider
+- MonoBahaviour
+- Quaternion
+- Object
+- Rigidbody
+- Scene
+- ScriptableObject
+- Sprite
+- Texture
+
 ## [1.1.O] - 2023-08-07
 ### Added
 - Plugins folder for package development